% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareRSSvsSRS.R
\name{compareRSSvsSRS}
\alias{compareRSSvsSRS}
\title{Compare RSS and SRS GWAS}
\usage{
compareRSSvsSRS(
  data,
  phenotype = "Yield",
  ranking = "PlantHeight",
  set.size = 5
)
}
\arguments{
\item{data}{Dataset.}

\item{phenotype}{Phenotype variable.}

\item{ranking}{Ranking variable.}

\item{set.size}{RSS set size.}
}
\value{
A list of GWAS results.
}
\description{
Compare top SNPs identified using RSS and SRS.
}
\examples{
data(gwas_data)
comp <- compareRSSvsSRS(gwas_data)

}
