% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssManhattan.R
\name{rssManhattan}
\alias{rssManhattan}
\title{Manhattan Plot}
\usage{
rssManhattan(results)
}
\arguments{
\item{results}{Output from standardGWAS() or rssGWAS().}
}
\value{
Manhattan plot.
}
\description{
Create Manhattan plot from GWAS results.
}
\examples{
data(gwas_data)
res <- standardGWAS(gwas_data)
rssManhattan(res)

}
