% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateGWAS.R
\name{simulateGWAS}
\alias{simulateGWAS}
\title{Simulate GWAS Data}
\usage{
simulateGWAS(n = 500, nSNP = 100, seed = 123)
}
\arguments{
\item{n}{Number of individuals.}

\item{nSNP}{Number of SNP markers.}

\item{seed}{Random seed.}
}
\value{
A data frame containing simulated SNP markers,
phenotype values and ranking variable.
}
\description{
Generate simulated GWAS genotype and phenotype data.
}
\examples{
sim <- simulateGWAS()
head(sim)

}
