% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srsSample.R
\name{srsSample}
\alias{srsSample}
\title{Simple Random Sampling}
\usage{
srsSample(data, n)
}
\arguments{
\item{data}{Data frame.}

\item{n}{Sample size.}
}
\value{
A sampled data frame.
}
\description{
Draw a simple random sample.
}
